Revert "workaround spring.components lost file with eclipse"
This reverts commit 05587acf07a7f242286e0c7662b8042caec8b2e4.

This fix is problematic as spring.components and spring-autoconfigure-metadata.properties
are not put in jar file if they are not added in classes/ (as source
folder is not meant to be added to jar file).

We rollback this change, and we must check that eclipse continues to
work (a fast check show that there is no problem with this revert;
unexpected, but we hope to be lucky).