mock-server: add not-bundle dependencies
Bundles are an issue as they package slf4j API and jul binding so it
conflicts with project slf4j log4j2 binding.

Bundles are kept in dependency management, but not-bundle are added to
allow project to migrate to these dependencies.

When multiple bindings are available, slf4j behavior is not
deterministic.

https://www.slf4j.org/codes.html#multiple_bindings